some refactor over entities collection

+++ kotiki-engine/include/kotiki-engine/utils/types.h
@@ -1,3 +1,3 @@
 #pragma once
 
-using coord_t = double;
+using coord_t = float;
